+++ webpack.config.js
@@ -146,6 +146,6 @@ if (process.env.NODE_ENV === "production") {
   ])
 } else {
   module.exports.plugins = (module.exports.plugins || []).concat([
-    new BundleAnalyzerPlugin(),
+    // new BundleAnalyzerPlugin(),
   ]);
 }